requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
apscheduler==3.10.4
httpx==0.27.2
jinja2==3.1.4
python-multipart==0.0.12
google-generativeai==0.8.2
pytz==2024.2
email-validator==2.2.0
python-dotenv==1.0.1

